Extract plugin cache to separate file
